requirements.txt
aiogram==3.5.0
pydub
requests
ffmpeg-python